Dev order action, and fix request sending logic.
First try filling and sending the request. If it succeeds, write the flow msg and register to mapper, or add the request back to queue and sleep for some time. So a request stays in queue until it is sent successfully. OrderRef is generated by active request via method provided by order provider so active request definitely has the order ref.